refactor(gateway): Q11.a — extract rest_v1/tasks/task_lists.rs (GAR-635)
Split the 4236-LOC tasks.rs monolith: task-list CRUD handlers and DTOs
(create_task_list, list_task_lists, get_task_list, patch_task_list,
delete_task_list + TaskListRow/CreateTaskListRequest/PatchTaskListRequest/
TaskListResponse/TaskListSummary/ListTaskListsResponse/ListTaskListsQuery +
ALLOWED_LIST_TYPES) extracted to rest_v1/tasks/task_lists.rs.

tasks.rs becomes tasks/mod.rs (~3572 LOC). Re-exports in mod.rs keep all
call-sites in router.rs and openapi.rs unchanged (openapi.rs updated to
reference super::tasks::task_lists::{...} for utoipa path resolution).

Zero behaviour change. cargo check + tests + clippy green.
