+++ crates/garraia-gateway/src/rest_v1/openapi.rs
@@ -116,11 +116,11 @@ impl Modify for SecurityAddon {
         super::memory::patch_memory,
         super::memory::pin_memory,
         super::memory::unpin_memory,
-        super::tasks::create_task_list,
-        super::tasks::list_task_lists,
-        super::tasks::get_task_list,
-        super::tasks::patch_task_list,
-        super::tasks::delete_task_list,
+        super::tasks::task_lists::create_task_list,
+        super::tasks::task_lists::list_task_lists,
+        super::tasks::task_lists::get_task_list,
+        super::tasks::task_lists::patch_task_list,
+        super::tasks::task_lists::delete_task_list,
         super::tasks::create_task,
         super::tasks::list_tasks,
         super::tasks::get_task,
